+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2019-11-21
+
 ### Changed
 
 - Extract Variable now extracts JSX Attributes (e.g. `<Header title="Home" />` will extract `"Home"` if cursor is on it).
@@ -489,7 +491,8 @@ function doSomething(someData) {
 - **[New refactoring]** Extract Variable
 - **[New refactoring]** Rename Symbol
 
-[unreleased]: https://github.com/nicoespeon/abracadabra/compare/1.0.1...HEAD
+[unreleased]: https://github.com/nicoespeon/abracadabra/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/nicoespeon/abracadabra/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/nicoespeon/abracadabra/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/nicoespeon/abracadabra/compare/0.11.0...1.0.0
 [0.11.0]: https://github.com/nicoespeon/abracadabra/compare/0.10.0...0.11.0